Refactor QueryResolver
